+++ frontend/src/app/layout.tsx
@@ -1,7 +1,8 @@
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 import "./globals.css";
-import Link from 'next/link'
+import Link from "next/link";
+import CartProviderWrapper from "@/components/CartProviderWrapper";
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
@@ -20,9 +21,9 @@ export const metadata: Metadata = {
 
 export default function RootLayout({
   children,
-}: Readonly<{
+}: {
   children: React.ReactNode;
-}>) {
+}) {
   return (
     <html lang="ko">
       <body
@@ -36,25 +37,33 @@ export default function RootLayout({
               width={100}
               height={100}
               className="pl-10"
-              
             />
           </Link>
           <Link href="/">
             <p className="pl-15 text-[#F9F9F2] text-3xl font-bold">STARBU9</p>
           </Link>
           <ul className="menu flex justify-between bg-[#005034] text-white ml-30 h-full items-center">
-            <Link href="/products" className="h-full hover:bg-[#377D22] w-[150] flex items-center justify-center">
+            <Link
+              href="/products"
+              className="h-full hover:bg-[#377D22] w-[150] flex items-center justify-center"
+            >
               <p className="text-xl">상품</p>
             </Link>
-            <Link href="/cart" className="h-full hover:bg-[#377D22] w-[150] flex items-center justify-center">
+            <Link
+              href="/cart"
+              className="h-full hover:bg-[#377D22] w-[150] flex items-center justify-center"
+            >
               <p className="text-xl">장바구니</p>
-              </Link>
-            <Link href="/order/list"  className="h-full hover:bg-[#377D22] w-[150] flex items-center justify-center">
+            </Link>
+            <Link
+              href="/order/list"
+              className="h-full hover:bg-[#377D22] w-[150] flex items-center justify-center"
+            >
               <p className="text-xl">주문</p>
             </Link>
           </ul>
         </header>
-        {children}
+        <CartProviderWrapper>{children}</CartProviderWrapper>
       </body>
     </html>
   );
