+++ frontend/src/app/products/page.tsx
@@ -1,8 +1,8 @@
 "use client";
 import React, { useEffect, useState } from "react";
 import { Product } from "../../lib/type/product";
-import ProductList from "@/commponents/ProductList";
-import ProductInfo from "@/commponents/ProductInfo";
+import ProductList from "@/components/ProductList";
+import ProductInfo from "@/components/ProductInfo";
 
 export default function Page() {
   const [products, setProducts] = useState<Product[]>([]); //상품 목록
