+++ frontend/src/components/ShoppingCart.tsx
